Push image to GHCR (in addition to Docker Hub)
This commit adds supports for multiple registries in the docker
workflow; built images are pushed to GitHub Container Registry (GHCR) in
addition to the standard Docker Hub.
GHCR is easier to use with a GitHub repository (we do not need to setup
a specific token), and displays built packages directly on the
repository page. It is thus a nice addition to Docker Hub, which is more
standard and well-known.