Improve tags in Makefile
This commit adds the branch-name and a short SHA to image tags,
so that we can differentiate between branches, and between commits in
the same branch. This allows comparing images easily (when locally
testing).
(Note that this is only relevant for the Makefile; the GitHub workflow
that pushes to Docker Hub uses different tagging rules).