Use dpkg multi-arch to install pgAdmin4 for amd64 in Dockerfile
pgAdmin4 is only available for the amd64 architecture. If the container
runs on arm64 architecture, by default dpkg (and apt) will refuse to
install pgAdmin4, because the package is not compatible. We can tell
dpkg to use the amd64 architecture for pgAdmin4, in order to force the
installation.

See https://wiki.debian.org/Multiarch/HOWTO

Surprisingly, pgadmin4-web cannot be installed, even with this
technique, but pgadmin4-server can. This may be because it cannot find
the pgadmin4-server:amd64 when it is a transitive dependency, yet finds
it when it is manually requested. This should not change much, because
pgadmin4-web only adds a few files and 1 dependency for Apache2-based
deployments.

CURRENTLY NOT WORKING!
Running pgadmin4 in WSGI mode does not work, because the psycopg lib
(used to query Postgres) does not find the `__impl__` attribute in its
implementation sub-module (psycopg_c)??? Other implementations
(binary/python) also fail for various reasons (cannot find the libpq.so
although it exists; cannot find a cryptography .so dll even though the
file exists...). Manually importing the psycopg in a Python REPL works,
but pgadmin4 still does not respond, which is very strange.