Install pgAdmin4 through pip to support amd64/arm64 platforms
The pgAdmin4 Debian packages are only available for the amd64 platform.
This covers the majority of users, but not those with an arm64 platform,
especially the recent Apple Silicon processors.

Our previous version required to emulate the container (i.e., the
container assumed an amd64 platform, and the host had to emulate if it
was different).

This commit changes the way we install pgAdmin4, to use the Python
package (through pip) instead of the Debian package. This package has no
dependency on the platform (although it works a bit differently) and
thus will allow us to build containers with support for either platform.

- The Dockerfile now installs python3, pip3, and pgAdmin4 through pip
  rather than relying on apt/dpkg. Note that deployment using `uwsgi` do
  not seem to work anymore; we have thus removed `uwsgi` from the list of
  installed packages.
- The pgadmin_config.py file now specifies the host to bind, because we
  directly rely on pgAdmin4 itself (using Flask) to serve the Web App,
  instead of uwsgi as previously.
- It also specifies the path to Postgres' binaries, because the Python
  package's config_distro.py does not set them. (And we want to minimize
  the configuration that the user must do)
- The run_pgadmin.sh does not invoke uwsgi anymore, as stated earlier.
  We instead directly invoke pgAdmin4, which uses Flask internally.

Note: there is a slight bug when trying to restore a database in this
  version; selecting a file through the file chooser appends a `None:`
  at the beginning (which makes the restoration fail). Setting the path
  manually (without `None:`) works perfectly.