Detect the installation path of pgadmin4
We previously hard-coded the path to which pgadmin4 is installed (as a
Python package, e.g., in `/usr/local/lib/python3.11/dist-packages/`).
The problem is that, if the base image updates its version of Python
(or if some reason pgadmin4 is no longer installed to this exact path),
our assumption fails and we cannot build the image (or worse, build an
image that does not work).

This commit replaces the hard-coded with a call to Python to find the
installation path. We assume that pgadmin4 will have a single path,
but that should be an acceptable assumption (or at least that the "main"
installation folder will be specified as the first path).